Struct oration::config::Config

pub struct Config {
    pub host: String,
    pub blog_name: String,
    pub salt: String,
    pub author: Author,
    pub nesting_limit: u32,
    pub edit_timeout: f32,
    pub notifications: Notifications,
    pub telegram: Telegram,
}

The main struct which all input data from oration.yaml is pushed into.

Fields
host: String

Top level location of the blog we are serving.

blog_name: String

Name of the blog we are serving.

salt: String

A salt for slightly more anonymous anonymous user identification.

author: Author

Blog Author to highlight as an authority in comments.

nesting_limit: u32

Limit of thread nesting in comments.

edit_timeout: f32

Time limit that restricts user editing of their own comments.

notifications: Notifications

Email notification system and connection details.

telegram: Telegram

Telegram notification endpoint details.

Methods

impl Config

pub fn load() -> Result<Config, Error>

Reads and parses data from the oration.yaml file and command line arguments.

fn parse(&self) -> Result<(), Error>

Additional checks to the configuration file that cannot be done implicitly by the type checker.