Function oration::edit_comment

fn edit_comment(
    conn: Conn, 
    config: State<Config>, 
    identifier: CommentId, 
    hash: AuthHash, 
    edits: Result<Form<FormEdit>, Option<String>>, 
    remote_addr: SocketAddr
) -> Result<Json<CommentEdits>, Failure>

Requests an update to a comment from a user, which may or may not occur based on a number of possibilities: authentication issues, over time, etc.