Struct oration::models::comments::NewComment

struct NewComment<'c> {
    tid: i32,
    parent: Option<i32>,
    created: NaiveDateTime,
    modified: Option<NaiveDateTime>,
    mode: i32,
    remote_addr: Option<&'c str>,
    text: &'c str,
    author: Option<String>,
    email: Option<String>,
    website: Option<String>,
    hash: String,
    likes: Option<i32>,
    dislikes: Option<i32>,
    voters: Option<Vec<u8>>,
}

Insertable reference to the comments table.

Fields
tid: i32

Reference to Thread.

parent: Option<i32>

Parent comment.

created: NaiveDateTime

Timestamp of creation.

modified: Option<NaiveDateTime>

Date modified it that's happened.

mode: i32

If the comment is live or under review. By default an active comment has mode 0. If the admin has reviews turned on, all new comments will be flagged as mode 1, or will be set with a default mode 0 if this feature is not enabled. A comment with mode 2 indicates this comment is deleted, although it contains responses below it. The deleted comment with therefore be handled differently.

remote_addr: Option<&'c str>

Remote IP.

text: &'c str

Actual comment.

author: Option<String>

Commentors author if given.

email: Option<String>

Commentors email address if given.

website: Option<String>

Commentors website if given.

hash: String

Sha224 hash to identify commentor.

likes: Option<i32>

Number of likes a comment has recieved.

dislikes: Option<i32>

Number of dislikes a comment has recieved.

voters: Option<Vec<u8>>

Who are the voters on this comment.