Struct oration::models::comments::PrintedComment

struct PrintedComment {
    id: i32,
    parent: Option<i32>,
    text: String,
    author: Option<String>,
    email: Option<String>,
    url: Option<String>,
    hash: String,
    created: NaiveDateTime,
    likes: Option<i32>,
    dislikes: Option<i32>,
}

Subset of the comments table which is to be sent to the frontend.

Fields
id: i32

Primary key.

parent: Option<i32>

Parent comment.

text: String

Actual comment.

author: Option<String>

Commentors author if given.

email: Option<String>

Commentors email address if given.

url: Option<String>

Commentors website if given.

hash: String

Commentors indentifier.

created: NaiveDateTime

Timestamp of creation.

likes: Option<i32>

Number of likes a comment has recieved.

dislikes: Option<i32>

Number of dislikes a comment has recieved.

Methods

impl PrintedComment

fn list(
conn: &SqliteConnection,
path: &str
) -> Result<Vec<PrintedComment>, Error>

Returns a list of all comments for a given post denoted via the path variable.

pub fn get(conn: &SqliteConnection, id: i32) -> Result<PrintedComment, Error>

Returns a comment based on its' unique ID.