Trait oration::models::preferences::_impl_identifiable_for_preference::diesel::query_source::QuerySource

pub trait QuerySource {
    type FromClause;
    type DefaultSelection: SelectableExpression<Self>;
    fn from_clause(&self) -> Self::FromClause;

    fn default_selection(&self) -> Self::DefaultSelection;
}

Represents a type which can appear in the FROM clause. Apps should not need to concern themselves with this trait.

Types which implement this trait include:

Tables generated by the table! macro

Internal structs which represent joins

A select statement which has had no query builder methods called on it, other than those which can affect the from clause.

Associated Types

type FromClause

The type returned by from_clause

type DefaultSelection: SelectableExpression<Self>

The type returned by default_selection

Required Methods

fn from_clause(&self) -> Self::FromClause

The actual FROM clause of this type. This is typically only called in QueryFragment implementations.

fn default_selection(&self) -> Self::DefaultSelection

The default select clause of this type, which should be used if no select clause was explicitly specified. This should always be a tuple of all the desired columns, not star