Macro oration::models::preferences::_impl_queryable_for_preference::diesel::impl_selectable_expression

macro_rules! impl_selectable_expression {
    ( $ struct_name : ident ) => { ... };
    ( $ struct_name : ident < $ ( $ ty_params : ident ) , + > ) => { ... };
    (
ty_params = ( $ ( $ ty_params : ident ) , * ) , struct_ty = $ struct_ty : ty ,
) => { ... };
}

This will implement SelectableExpression and AppearsOnTable for "simple" composite nodes where the where clause is roughly AllTyParams: SelectableExpression<QS>, Self: Expression.

This macro is exported because we want to be able to call it from other macros that are exported, but it is not part of our public API.