Struct oration::models::preferences::Preference

pub struct Preference {
    pub key: String,
    pub value: String,
}

Queryable, Identifiable reference to the preferences table.

Fields
key: String

Key

value: String

Value

Methods

impl Preference

pub fn set_session(conn: &SqliteConnection) -> Result<bool, Error>

Updates the sesssion key into the database only if the key does not exist. A default value is set in the migration schema and no other functions operate on this entry, so that should cover all bases.

pub fn get_session(conn: &SqliteConnection) -> Result<String, Error>

Returns the current session value from the database.