Module oration::models::threads

Threads table.

Modules

_impl_insertable_for_newthread

_impl_queryable_for_thread

Structs

NewThread
Insertable reference to the threads table.

Thread
Queryable reference to the threads table.

Constants

_IMPL_SERIALIZE_FOR_Thread

Functions

create
Saves a new thread for URI into the database. Returns the id of the new record.

gen_or_get_id
Returns a thread ID given creation details about it. If the thread exists, an ID is returned directly, otherwise an entry is created for it first

get_id
Returns the id of a thread from the database for a given URI.

verify_post
Checks that the path posted actually exists on the host. Should minimise the injection attack surface.