Macro oration::schema::comments::columns::_impl_query_id_for_author::diesel::__diesel_parse_type_args

macro_rules! __diesel_parse_type_args {
    (
data = $ data : tt , callback = $ callback : ident , tokens = $ tokens : tt ,
) => { ... };
    (
data = $ data : tt , callback = $ callback : ident , args = (
$ ( $ args : tt ) * ) , bounds = ( $ ( $ bounds : tt ) * ) , tokens = (
$ next_arg : ident : $ ( $ tokens : tt ) * ) , ) => { ... };
    (
data = $ data : tt , callback = $ callback : ident , args = (
$ ( $ args : tt ) * ) , bounds = ( $ ( $ bounds : tt ) * ) , tokens = (
$ next_arg : ident $ ( $ tokens : tt ) * ) , ) => { ... };
    (
brackets = (  ) , data = $ data : tt , callback = $ callback : ident , args =
$ args : tt , bounds = ( $ ( $ bounds : tt ) * ) , tokens = (
> $ ( $ tokens : tt ) * ) , ) => { ... };
    (
brackets = (  ) , data = $ data : tt , callback = $ callback : ident , args =
$ args : tt , bounds = ( $ ( $ bounds : tt ) * ) , tokens = (
, $ ( $ tokens : tt ) * ) , ) => { ... };
    (
brackets = ( $ ( $ brackets : tt ) * ) , data = $ data : tt , callback = $
callback : ident , args = $ args : tt , bounds = ( $ ( $ bounds : tt ) * ) ,
tokens = ( < $ ( $ tokens : tt ) * ) , ) => { ... };
    (
brackets = ( < $ ( $ brackets : tt ) * ) , data = $ data : tt , callback = $
callback : ident , args = $ args : tt , bounds = ( $ ( $ bounds : tt ) * ) ,
tokens = ( >> $ ( $ tokens : tt ) * ) , ) => { ... };
    (
brackets = ( < $ ( $ brackets : tt ) * ) , data = $ data : tt , callback = $
callback : ident , args = $ args : tt , bounds = ( $ ( $ bounds : tt ) * ) ,
tokens = ( > $ ( $ tokens : tt ) * ) , ) => { ... };
    (
brackets = $ brackets : tt , data = $ data : tt , callback = $ callback :
ident , args = $ args : tt , bounds = ( $ ( $ bounds : tt ) * ) , tokens = (
$ token : tt $ ( $ tokens : tt ) * ) , ) => { ... };
    (
data = $ data : tt , callback = $ callback : ident , args = $ args : tt ,
bounds = $ bounds : tt , tokens = ( , $ ( $ tokens : tt ) * ) , ) => { ... };
    (
data = $ data : tt , callback = $ callback : ident , args = $ args : tt ,
bounds = $ bounds : tt , tokens = ( > $ ( $ tokens : tt ) * ) , ) => { ... };
}

Parses a sequence of type parameters and their bounds.

Ideally we would just be able to write this as $($param:ident $(: $bound:ty)*),* $(,)*, but Rust doesn't allow a ty fragment in a trait bound position.

Assuming we don't care about lifetimes or existentials, we could also possibly write $($param:ident $(: $bound:path)+*),* $(,)* but Rust doesn't allow + as a separator. In fact, we can't even use the path fragment at all, as Rust does not allow it to be followed by + or tt.

This macro takes three arguments.

data: Any arbitrary token tree you'd like given back to you when parsing is complete.

callback: The name of the macro to call when parsing is complete

tokens: The tokens to be parsed.

This macro will consume tokens until it reaches a > that was not preceded by a <. It will then invoke callback with 4 arguments:

data: Whatever you gave this macro.

type_args: The names of the type parameters. Always contains a trailing comma if non-empty.

type_args_with_bounds: The arguments and their bounds. Always contains a trailing comma if non-empty

unparsed_tokens: Any tokens that were not consumed by this macro.