Module oration::schema::comments::columns::_impl_query_id_for_created::diesel::query_builder::functions

Functions

delete
Creates a DELETE statement.

insert_into
Creates an INSERT statement for the target table.

insert_or_ignore_into
Creates an INSERT [OR] IGNORE statement.

replace_into
Creates a REPLACE statement.

select
Creates a bare select statement, with no from clause. Primarily used for testing diesel itself, but likely useful for third party crates as well. The given expressions must be selectable from anywhere.

sql_query
Construct a full SQL query using raw SQL.

update
Creates an UPDATE statement.