Trait oration::schema::comments::columns::_impl_query_id_for_created::diesel::query_builder::IntoUpdateTarget

pub trait IntoUpdateTarget: HasTable {
    type WhereClause;
    fn into_update_target(self) -> UpdateTarget<Self::Table, Self::WhereClause>;
}

A type which can be passed to update or delete.

Apps will never need to implement this type directly. There are three kinds which implement this trait. Tables, queries which have only had filter called on them, and types which implement Identifiable.

When a table is passed to update, every row in the table will be updated. You can scope this down by calling filter which will result in UPDATE your_table SET ... WHERE args_to_filter. Passing a type which implements Identifiable is the same as passing SomeStruct::table().find(some_struct).

Associated Types

type WhereClause

What is the WHERE clause of this target?

Required Methods

fn into_update_target(self) -> UpdateTarget<Self::Table, Self::WhereClause>

Decomposes self into the table and where clause.