Trait oration::schema::comments::columns::_impl_query_id_for_dislikes::diesel::row::Row

pub trait Row<DB> where
    DB: Backend, {
    fn take(&mut self) -> Option<&<DB as Backend>::RawValue>;

    fn next_is_null(&self, count: usize) -> bool;

    fn advance(&mut self, count: usize) { ... }
}

Represents a single database row. Apps should not need to concern themselves with this trait.

This trait is only used as an argument to FromSqlRow.

Required Methods

fn take(&mut self) -> Option<&<DB as Backend>::RawValue>

Returns the value of the next column in the row.

fn next_is_null(&self, count: usize) -> bool

Returns whether the next count columns are all NULL.

If this method returns true, then the next count calls to take would all return None.

Provided Methods

fn advance(&mut self, count: usize)

Skips the next count columns. This method must be called if you are choosing not to call take as a result of next_is_null returning true.