Enum oration::schema::comments::columns::_impl_query_id_for_hash::diesel::result::DatabaseErrorKind

pub enum DatabaseErrorKind {
    UniqueViolation,
    ForeignKeyViolation,
    UnableToSendCommand,
    __Unknown,
}

The kind of database error that occurred.

This is not meant to exhaustively cover all possible errors, but is used to identify errors which are commonly recovered from programmatically. This enum is not intended to be exhaustively matched, and new variants may be added in the future without a major version bump.

Variants
UniqueViolation

A unique constraint was violated.

ForeignKeyViolation

A foreign key constraint was violated.

UnableToSendCommand

The query could not be sent to the database due to a protocol violation.

An example of a case where this would occur is if you attempted to send a query with more than 65000 bind parameters using PostgreSQL.

__Unknown