Module oration::schema::comments::columns::_impl_query_id_for_id::diesel::expression_methods

Adds various methods to construct new expressions. These traits are exported by default, and implemented automatically.

You can rely on the methods provided by this trait existing on any Expression of the appropriate type. You should not rely on the specific traits existing, their names, or their organization.

Traits

BoolExpressionMethods
Methods present on boolean expressions

EqAll
This method is used by FindDsl to work with tuples. Because we cannot express this without specialization or overlapping impls, it is brute force implemented on columns in the column! macro.

EscapeExpressionMethods
Adds the escape method to LIKE and NOT LIKE. This is used to specify the escape character for the pattern.

ExpressionMethods
Methods present on all expressions, except tuples

NullableExpressionMethods
Methods present on all expressions

TextExpressionMethods
Methods present on text expressions