Struct oration::schema::comments::columns::_impl_query_id_for_mode::diesel::query_builder::AstPass

pub struct AstPass<'a, DB> where
    DB: Backend,
    <DB as Backend>::QueryBuilder: 'a,
    <DB as Backend>::BindCollector: 'a,
    <DB as TypeMetadata>::MetadataLookup: 'a,  {
    internals: AstPassInternals<'a, DB>,
}

The primary type used when walking a Diesel AST during query execution.

Executing a query is generally done in multiple passes. This list includes, but is not limited to:

Generating the SQL

Collecting and serializing bound values (sent separately from the SQL)

Determining if a query is safe to store in the prepared statement cache

When adding a new type that is used in a Diesel AST, you don't need to care about which specific passes are being performed, nor is there any way for you to find out what the current pass is. You should simply call the relevant methods and trust that they will be a no-op if they're not relevant to the current pass.

Fields
internals: AstPassInternals<'a, DB>