Trait oration::schema::comments::columns::_impl_query_id_for_modified::diesel::connection::Connection

pub trait Connection: SimpleConnection + Send {
    type Backend: Backend;
    type TransactionManager: TransactionManager<Self>;
    fn establish(database_url: &str) -> Result<Self, ConnectionError>;

    fn execute(&self, query: &str) -> Result<usize, Error>;

    fn query_by_index<T, U>(&self, source: T) -> Result<Vec<U>, Error>
    where
        T: AsQuery,
        U: Queryable<<T as AsQuery>::SqlType, Self::Backend>,
        <T as AsQuery>::Query: QueryFragment<Self::Backend>,
        <T as AsQuery>::Query: QueryId,
        Self::Backend: HasSqlType<<T as AsQuery>::SqlType>;

    fn query_by_name<T, U>(&self, source: &T) -> Result<Vec<U>, Error>
    where
        T: QueryFragment<Self::Backend> + QueryId,
        U: QueryableByName<Self::Backend>;

    fn execute_returning_count<T>(&self, source: &T) -> Result<usize, Error>
    where
        T: QueryFragment<Self::Backend> + QueryId;

    fn transaction_manager(&self) -> &Self::TransactionManager;

    fn transaction<T, E, F>(&self, f: F) -> Result<T, E>
    where
        E: From<Error>,
        F: FnOnce() -> Result<T, E>,
    { ... }

    fn begin_test_transaction(&self) -> Result<(), Error> { ... }

    fn test_transaction<T, E, F>(&self, f: F) -> T
    where
        E: Debug,
        F: FnOnce() -> Result<T, E>,
    { ... }
}

A connection to a database

Associated Types

type Backend: Backend

The backend this type connects to

type TransactionManager: TransactionManager<Self>

Required Methods

fn establish(database_url: &str) -> Result<Self, ConnectionError>

Establishes a new connection to the database

The argument to this method varies by backend. See the documentation for that backend's connection class for details about what it accepts.

fn execute(&self, query: &str) -> Result<usize, Error>

fn query_by_index<T, U>(&self, source: T) -> Result<Vec<U>, Error> where
T: AsQuery,
U: Queryable<<T as AsQuery>::SqlType, Self::Backend>,
<T as AsQuery>::Query: QueryFragment<Self::Backend>,
<T as AsQuery>::Query: QueryId,
Self::Backend: HasSqlType<<T as AsQuery>::SqlType>,

fn query_by_name<T, U>(&self, source: &T) -> Result<Vec<U>, Error> where
T: QueryFragment<Self::Backend> + QueryId,
U: QueryableByName<Self::Backend>,

fn execute_returning_count<T>(&self, source: &T) -> Result<usize, Error> where
T: QueryFragment<Self::Backend> + QueryId,

fn transaction_manager(&self) -> &Self::TransactionManager

Provided Methods

fn transaction<T, E, F>(&self, f: F) -> Result<T, E> where
E: From<Error>,
F: FnOnce() -> Result<T, E>,

Executes the given function inside of a database transaction

If there is already an open transaction, savepoints will be used instead.

Example

use diesel::result::Error;

conn.transaction::<_, Error, _>(|| {
    diesel::insert_into(users)
        .values(name.eq("Ruby"))
        .execute(&conn)?;

    let all_names = users.select(name).load::<String>(&conn)?;
    assert_eq!(vec!["Sean", "Tess", "Ruby"], all_names);

    Ok(())
})?;

conn.transaction::<(), _, _>(|| {
    diesel::insert_into(users)
        .values(name.eq("Pascal"))
        .execute(&conn)?;

    let all_names = users.select(name).load::<String>(&conn)?;
    assert_eq!(vec!["Sean", "Tess", "Ruby", "Pascal"], all_names);

    // If we want to roll back the transaction, but don't have an
    // actual error to return, we can return `RollbackTransaction`.
    Err(Error::RollbackTransaction)
});

let all_names = users.select(name).load::<String>(&conn)?;
assert_eq!(vec!["Sean", "Tess", "Ruby"], all_names);

fn begin_test_transaction(&self) -> Result<(), Error>

Creates a transaction that will never be committed. This is useful for tests. Panics if called while inside of a transaction.

fn test_transaction<T, E, F>(&self, f: F) -> T where
E: Debug,
F: FnOnce() -> Result<T, E>,

Executes the given function inside a transaction, but does not commit it. Panics if the given function returns an error.

Example

use diesel::result::Error;

conn.test_transaction::<_, Error, _>(|| {
    diesel::insert_into(users)
        .values(name.eq("Ruby"))
        .execute(&conn)?;

    let all_names = users.select(name).load::<String>(&conn)?;
    assert_eq!(vec!["Sean", "Tess", "Ruby"], all_names);

    Ok(())
});

// Even though we returned `Ok`, the transaction wasn't committed.
let all_names = users.select(name).load::<String>(&conn)?;
assert_eq!(vec!["Sean", "Tess"], all_names);