Module oration::schema::comments::columns::_impl_query_id_for_modified::diesel::expression

AST types representing various typed SQL expressions.

Almost all types implement either Expression or AsExpression.

The most common expression to work with is a Column. There are various methods that you can call on these, found in expression_methods.

You can also use numeric operators such as + on expressions of the appropriate type.

Any primitive which implements ToSql will also implement AsExpression, allowing it to be used as an argument to any of the methods described here.

Modules

array_comparison

bound

coerce

count

dsl

exists

functions

grouped

helper_types
The types in this module are all shorthand for PredicateType<Lhs, AsExpr<Rhs, Lhs>>. Since we often need to return concrete types, instead of a boxed trait object, these can be useful for writing concise return types.

nullable

operators

ops

sql_literal

subselect

Traits

AppearsOnTable
Indicates that all elements of an expression are valid given a from clause.

AsExpression
Converts a type to its representation for use in Diesel's query builder.

AsExpressionList
Converts a tuple of values into a tuple of Diesel expressions.

BoxableExpression
Helper trait used when boxing expressions.

Expression
Represents a typed fragment of SQL.

IntoSql
Converts a type to its representation for use in Diesel's query builder.

NonAggregate
Marker trait to indicate that an expression does not include any aggregate functions.

SelectableExpression
Indicates that an expression can be selected from a source.