Module oration::schema::comments::columns::_impl_query_id_for_parent::diesel

Modules

associations
Traits related to relationships between multiple tables.

backend
Types which represent various database backends

connection
Types related to database connections

data_types
Structs to represent the primitive equivalent of SQL types where there is no existing Rust primitive, or where using it would be confusing (such as date and time types). This module will re-export all backend specific data structures when compiled against that backend.

deserialize
Types and traits related to deserializing values from the database

dsl
Includes various helper types and bare functions which are named too generically to be included in prelude, but are often used when using Diesel.

expression
AST types representing various typed SQL expressions.

expression_methods
Adds various methods to construct new expressions. These traits are exported by default, and implemented automatically.

helper_types
Provide helper types for concisely writing the return type of functions. As with iterators, it is unfortunately difficult to return a partially constructed query without exposing the exact implementation of the function. Without higher kinded types, these various DSLs can't be combined into a single trait for boxing purposes.

insertable

migration
Representation of migrations

prelude
Re-exports important traits and types. Meant to be glob imported when using Diesel.

query_builder
Contains traits responsible for the actual construction of SQL statements

query_dsl
Traits that construct SELECT statements

query_source
Types related to describing schema, and interactions between tables.

result
Errors, type aliases, and functions related to working with Result.

row
Contains the Row trait

serialize
Types and traits related to serializing values for the database

sql_types
Types which represent a SQL data type.

sqlite
Provides types and functions related to working with SQLite

types 	 [
Deprecated
]

Macros

__diesel_column

__diesel_for_each_tuple

__diesel_generate_ops_impls_if_date_time

__diesel_generate_ops_impls_if_numeric

__diesel_operator_body

__diesel_operator_to_sql

__diesel_parse_type_args
Parses a sequence of type parameters and their bounds.

__diesel_sql_function_body

__diesel_sqlite_register_fn

__diesel_table_impl

__diesel_table_query_source_impl

__diesel_use_everything
Used by diesel_derives, which can't access $crate

allow_tables_to_appear_in_same_query
Allow two or more tables which are otherwise unrelated to be used together in a query.

date_time_expr

diesel_infix_operator
Useful for libraries adding support for new SQL types. Apps should never need to call this.

diesel_postfix_operator
Useful for libraries adding support for new SQL types. Apps should never need to call this.

diesel_prefix_operator
Useful for libraries adding support for new SQL types. Apps should never need to call this.

impl_query_id 	 [
Deprecated
]
Provides a standard implementation of QueryId. Apps should not need to concern themselves with this macro.

impl_selectable_expression
This will implement SelectableExpression and AppearsOnTable for "simple" composite nodes where the where clause is roughly AllTyParams: SelectableExpression<QS>, Self: Expression.

joinable
Allow two tables to be referenced in a join query without providing an explicit ON clause.

joinable_inner

no_arg_sql_function
Declare a 0 argument SQL function for use in your code. This will generate a unit struct, which is an expression representing calling this function. See now for example output. now was generated using:

no_arg_sql_function_body

no_arg_sql_function_body_except_to_sql

not_none
Gets the value out of an option, or returns an error.

numeric_expr
Indicates that an expression allows all numeric operators. If you create new SQL functions that return a numeric type, you should invoke this macro that type. Unfortunately, Rust disallows us from automatically implementing Add for types which implement Expression, under its orphan rules.

operator_allowed
Implements the Rust operator for a given type. If you create a new SQL function, which returns a type that you'd like to use an operator on, you should invoke this macro. Unfortunately, Rust disallows us from automatically implementing Add and other traits from std::ops, under its orphan rules.

sql_function
Declare a sql function for use in your code.

sql_function_body 	 [
Deprecated
]

static_cond

table
Specifies that a table exists, and what columns it has. This will create a new public module, with the same name, as the name of the table. In this module, you'll find a unit struct named table, and a unit struct with the names of each of the columns. In the definition, you can also specify an additional set of columns which exist, but should not be selected by default (for example, for things like full text search)

table_body