Struct oration::schema::comments::columns::_impl_query_id_for_parent::diesel::sqlite::SqliteConnection

pub struct SqliteConnection {
    statement_cache: StatementCache<Sqlite, Statement>,
    raw_connection: RawConnection,
    transaction_manager: AnsiTransactionManager,
}

Connections for the SQLite backend. Unlike other backends, "connection URLs" for SQLite are file paths or special identifiers like :memory.

Fields
statement_cache: StatementCache<Sqlite, Statement> raw_connection: RawConnection transaction_manager: AnsiTransactionManager
Methods

impl SqliteConnection

pub fn immediate_transaction<T, E, F>(&self, f: F) -> Result<T, E> where
E: From<Error>,
F: FnOnce() -> Result<T, E>,

Run a transaction with BEGIN IMMEDIATE

This method will return an error if a transaction is already open.

Example

conn.immediate_transaction(|| {
    // Do stuff in a transaction
    Ok(())
})

pub fn exclusive_transaction<T, E, F>(&self, f: F) -> Result<T, E> where
E: From<Error>,
F: FnOnce() -> Result<T, E>,

Run a transaction with BEGIN EXCLUSIVE

This method will return an error if a transaction is already open.

Example

conn.exclusive_transaction(|| {
    // Do stuff in a transaction
    Ok(())
})

pub fn register_sql_function<ArgsSqlType, RetSqlType, Args, Ret, F>(
&self,
fn_name: &str,
deterministic: bool,
f: F
) -> Result<(), Error> where
Args: Queryable<ArgsSqlType, Sqlite>,
F: FnMut(Args) -> Ret + Send + 'static,
Ret: ToSql<RetSqlType, Sqlite>,
Sqlite: HasSqlType<RetSqlType>,