Trait oration::schema::comments::columns::_impl_query_id_for_text::diesel::expression::AsExpressionList

pub trait AsExpressionList<ST> {
    type Expression;
    fn as_expression_list(self) -> Self::Expression;
}

Converts a tuple of values into a tuple of Diesel expressions.

This trait is similar to AsExpression, but it operates on tuples. The expressions must all be of the same SQL type.

Associated Types

type Expression

The final output expression

Required Methods

fn as_expression_list(self) -> Self::Expression

Perform the conversion