Trait oration::schema::comments::columns::_impl_query_id_for_tid::diesel::expression::IntoSql

pub trait IntoSql {
    fn into_sql<T>(self) -> Self::Expression
    where
        Self: AsExpression<T>,
    { ... }

    fn as_sql<'a, T>(&'a self) -> <&'a Self as AsExpression<T>>::Expression
    where
        &'a Self: AsExpression<T>,
    { ... }
}

Converts a type to its representation for use in Diesel's query builder.

This trait only exists to make usage of AsExpression more ergonomic when the SqlType cannot be inferred. It is generally used when you need to use a Rust value as the left hand side of an expression, or when you want to select a constant value.

Example

use diesel::sql_types::Text;
let names = users::table
    .select("The Amazing ".into_sql::<Text>().concat(users::name))
    .load(&conn);
let expected_names = vec![
    "The Amazing Sean".to_string(),
    "The Amazing Tess".to_string(),
];
assert_eq!(Ok(expected_names), names);

Provided Methods

fn into_sql<T>(self) -> Self::Expression where
Self: AsExpression<T>,

Convert self to an expression for Diesel's query builder.

There is no difference in behavior between x.into_sql::<Y>() and AsExpression::<Y>::as_expression(x).

Important traits for &'a mut W
impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

fn as_sql<'a, T>(&'a self) -> <&'a Self as AsExpression<T>>::Expression where
&'a Self: AsExpression<T>,

Convert &self to an expression for Diesel's query builder.

There is no difference in behavior between x.as_sql::<Y>() and AsExpression::<Y>::as_expression(&x).