Trait oration::schema::comments::columns::_impl_query_id_for_voters::diesel::query_builder::AsQuery

pub trait AsQuery where
    <Self::Query as Query>::SqlType == Self::SqlType, {
    type SqlType;
    type Query: Query;
    fn as_query(self) -> Self::Query;
}

Types that can be converted into a complete, typed SQL query.

This is used internally to automatically add the right select clause when none is specified, or to automatically add RETURNING * in certain contexts.

A type which implements this trait is guaranteed to be valid for execution.

Associated Types

type SqlType

The SQL type of Self::Query

type Query: Query

What kind of query does this type represent?

Required Methods

fn as_query(self) -> Self::Query

Converts a type which semantically represents a SQL query into the actual query being executed. See the trait level docs for more.