Module oration::schema::comments::columns::_impl_query_id_for_voters::diesel::result

Errors, type aliases, and functions related to working with Result.

Structs

UnexpectedNullError
An unexpected NULL was encountered during deserialization

Enums

ConnectionError
Errors which can occur during Connection::establish

DatabaseErrorKind
The kind of database error that occurred.

Error
Represents all the ways that a query can fail.

Traits

DatabaseErrorInformation
Information about an error that was returned by the database.

OptionalExtension
See the method documentation.

Type Definitions

ConnectionResult
A specialized result type for establishing connections.

QueryResult
A specialized result type for queries.