Trait oration::schema::comments::columns::_impl_query_id_for_website::diesel::deserialize::FromSqlRow

pub trait FromSqlRow<A, DB> where
    DB: Backend, {
    const FIELDS_NEEDED: usize;

    fn build_from_row<T>(
        row: &mut T
    ) -> Result<Self, Box<Error + 'static + Sync + Send>>
    where
        T: Row<DB>;
}

Deserialize one or more fields.

All types which implement FromSql should also implement this trait. This trait differs from FromSql in that it is also implemented by tuples. Implementations of this trait are usually derived.

In the future, we hope to be able to provide a blanket impl of this trait for all types which implement FromSql. However, as of Diesel 1.0, such an impl would conflict with our impl for tuples.

Deriving

This trait can be automatically derived by Diesel for any type which implements FromSql. There are no options or special considerations needed for this derive. Note that #[derive(FromSqlRow)] will also generate a Queryable implementation.

Associated Constants

const FIELDS_NEEDED: usize

FIELDS_NEEDED: usize = /// The number of fields that this type will consume. Must be equal to
/// the number of times you would call `row.take()` in `build_from_row`
const FIELDS_NEEDED: usize = 1;

The number of fields that this type will consume. Must be equal to the number of times you would call row.take() in build_from_row

Required Methods

fn build_from_row<T>(
row: &mut T
) -> Result<Self, Box<Error + 'static + Sync + Send>> where
T: Row<DB>,

See the trait documentation.