Module oration::schema::comments::columns::_impl_query_id_for_website::diesel::dsl

Includes various helper types and bare functions which are named too generically to be included in prelude, but are often used when using Diesel.

Type Definitions

Distinct
Represents the return type of .distinct()

Filter
Represents the return type of .filter(predicate)

Find
Represents the return type of .find(pk)

FindBy
Represents the return type of .filter(lhs.eq(rhs))

ForKeyShare
Represents the return type of .for_key_share()

ForNoKeyUpdate
Represents the return type of .for_no_key_update()

ForShare
Represents the return type of .for_share()

ForUpdate
Represents the return type of .for_update()

InnerJoin
Represents the return type of .inner_join(rhs)

IntoBoxed
Represents the return type of .into_boxed::<'a, DB>()

LeftJoin
Represents the return type of .left_join(rhs)

Limit
Represents the return type of .limit()

NoWait
Represents the return type of .no_wait()

Offset
Represents the return type of .offset()

OrFilter
Represents the return type of .or_filter(predicate)

Order
Represents the return type of .order(ordering)

Select
Represents the return type of .select(selection)

SingleValue
Represents the return type of .single_value()

SkipLocked
Represents the return type of .skip_locked()

ThenOrderBy
Represents the return type of .then_order_by(ordering)

Update
Represents the return type of update(lhs).set(rhs)