Struct oration::schema::comments::columns::_impl_query_id_for_website::diesel::query_source::joins::JoinOn

pub struct JoinOn<Join, On> {
    join: Join,
    on: On,
}

A query source representing the join between two tables with an explicit ON given. Join should usually be referenced instead, as all "type safety" traits are implemented in terms of Join implementing them.

Fields
join: Join on: On