Struct oration::schema::comments::columns::_impl_query_id_for_website::diesel::sql_types::Nullable

pub struct Nullable<ST>(ST)
 where
    ST: NotNull;

The nullable SQL type.

This wraps another SQL type to indicate that it can be null. By default all values are assumed to be NOT NULL.

ToSql impls

Any T which implements ToSql<ST>

Option<T> for any T which implements ToSql<ST>

FromSql impls

Option<T> for any T which implements FromSql<ST>