Trait oration::schema::preferences::_impl_query_id_for_table::diesel::expression::SelectableExpression

pub trait SelectableExpression<QS>: AppearsOnTable<QS> where
    QS: ?Sized, { }

Indicates that an expression can be selected from a source.

Columns will implement this for their table. Certain special types, like CountStar and Bound will implement this for all sources. Most compound expressions will implement this if each of their parts implement it.

Notably, columns will not implement this trait for the right side of a left join. To select a column or expression using a column from the right side of a left join, you must call .nullable() on it.