Module oration::schema::preferences::columns::_impl_query_id_for_key::diesel::query_source

Types related to describing schema, and interactions between tables.

Most traits in this module are derived or generated by table!.

Modules

joins

Structs

MoreThanOnce
A table appears in the from clause two or more times.

Never
A table never appears in the from clause.

Once
A table appears in the from clause exactly one time.

Traits

AppearsInFromClause
Determines how many times Self appears in QS

Column
A column on a database table. Types which implement this trait should have been generated by the table! macro.

Plus
Add two peano numbers together.

QuerySource
Represents a type which can appear in the FROM clause. Apps should not need to concern themselves with this trait.

Table
A SQL database table. Types which implement this trait should have been generated by the table! macro.