Module oration::schema::threads

Reexports

pub use self::columns::*;

Modules

columns
Contains all of the columns of this table

dsl
Re-exports all of the columns of this table, as well as the table struct renamed to the module name. This is meant to be glob imported for functions which only deal with one table.

Structs

table
The actual table struct

Constants

all_columns
A tuple of all of the columns on this table

Type Definitions

BoxedQuery
Helper type for representing a boxed query from this table

SqlType
The SQL type of all of the columns on this table