Module oration::schema::threads::columns::_impl_query_id_for_id::diesel::expression::helper_types

The types in this module are all shorthand for PredicateType<Lhs, AsExpr<Rhs, Lhs>>. Since we often need to return concrete types, instead of a boxed trait object, these can be useful for writing concise return types.

Type Definitions

And
The return type of lhs.and(rhs)

AsExpr
The type of Item when converted to an expression with the same type as TargetExpr

AsExprOf
The type of Item when converted to an expression of Type

Asc
The return type of expr.asc()

Between
The return type of lhs.between(lower, upper)

Desc
The return type of expr.desc()

Eq
The return type of lhs.eq(rhs)

EqAny
The return type of lhs.eq_any(rhs)

Escape
The return type of lhs.escape('x')

Gt
The return type of lhs.gt(rhs)

GtEq
The return type of lhs.ge(rhs)

IsNotNull
The return type of expr.is_not_null()

IsNull
The return type of expr.is_null()

Like
The return type of lhs.like(rhs)

Lt
The return type of lhs.lt(rhs)

LtEq
The return type of lhs.le(rhs)

NeAny
The return type of lhs.ne_any(rhs)

NotBetween
The return type of lhs.not_between(lower, upper)

NotEq
The return type of lhs.ne(rhs)

NotLike
The return type of lhs.not_like(rhs)

Nullable
The return type of expr.nullable()

Or
The return type of lhs.or(rhs)

SqlTypeOf
The SQL type of an expression