Module oration::schema::threads::columns::_impl_query_id_for_id::diesel::sqlite

Provides types and functions related to working with SQLite

Much of this module is re-exported from database agnostic locations. However, if you are writing code specifically to extend Diesel on SQLite, you may need to work with this module directly.

Modules

query_builder
The SQLite query builder

Structs

Sqlite
The SQLite backend

SqliteConnection
Connections for the SQLite backend. Unlike other backends, "connection URLs" for SQLite are file paths or special identifiers like :memory.

Enums

SqliteType
Determines how a bind parameter is given to SQLite