Struct oration::config::Config

pub struct Config {
    pub host: String,
}

Fields
host: String

Top level location of the blog we are serving.

Methods

impl Config

pub fn load() -> Result<Config, Error>

Reads and parses data from the oration.yaml file and command line arguments.

fn parse(&self) -> Result<(), Error>

Additional checks to the configuration file that cannot be done implicitly by the type checker.